+++ src/evox/algorithms/mo/hype.py
@@ -57,6 +57,7 @@ class HypE(Algorithm):
     """HypE algorithm
 
     link: https://direct.mit.edu/evco/article-abstract/19/1/45/1363/HypE-An-Algorithm-for-Fast-Hypervolume-Based-Many
+    Inspired by PlatEMO.
     """
 
     def __init__(
